requirements.txt
openai>=1.0.0      
python-dotenv       
pyyaml              
radon               
termcolor           
docker